Keep window geometry helper readable
The previous review fix preserved the Swift file budget by compressing the persisted-geometry helper, but that made the window sizing path harder to read. Move MainWindowHostingView beside CmuxMainWindow in the existing main-window source file, then restore the AppDelegate geometry helper to normal Swift formatting while keeping the file under budget.

Constraint: Swift file length budget must remain green

Rejected: Keep the semicolon-packed helper | non-idiomatic and called out by review

Confidence: high

Scope-risk: narrow

Tested: git diff --check

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Not-tested: XCTest per repository policy to prefer CI and user instruction forbidding direct xcodebuild
